+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Pending release
 
 #### Fixed
 - `Embed.clear()` not removing the timestamp.
+- `PartialInvite.inviter` error when missing inviter.
 
 ## 0.0.1-alpha
 Released on May 17, 2023.